requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==4.1.3
django-cors-headers==3.7.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
Markdown==3.3.4
numpy==1.22.3
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.6
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0